requirements.txt
# ──────────────────────────────────────
# igarateca – Backend Dependencies
# ──────────────────────────────────────

# Web framework
fastapi>=0.104.0,<1.0
uvicorn[standard]>=0.24.0,<1.0
python-dotenv>=1.0.0,<2.0
python-multipart>=0.0.6,<1.0

# Database / ORM
sqlalchemy>=2.0.0,<3.0
psycopg2-binary>=2.9.0,<3.0

# Auth
python-jose[cryptography]>=3.3.0,<4.0
passlib[bcrypt]>=1.7.4,<2.0
bcrypt>=4.0.0,<4.1

# Cache / Queue
redis>=5.0.0,<6.0

# Validation / Serialization
pydantic[email]>=2.0.0,<3.0

# Data / Export
pandas>=2.1.0,<3.0
openpyxl>=3.1.0,<4.0

# ETL – download
requests>=2.31.0,<3.0

# CORS (included with fastapi, listed for clarity)
# starlette – comes with fastapi
